utility methods should go to utils dir
