+++ cloud/pkg/admissioncontroller/utils/generic.go
@@ -1,0 +1,20 @@
+package utils
+
+import (
+	"net"
+)
+
+func GetIPv4Addr() string {
+	addrs, err := net.InterfaceAddrs()
+	if err != nil {
+		return ""
+	}
+	for _, address := range addrs {
+		if ipnet, ok := address.(*net.IPNet); ok && !ipnet.IP.IsLoopback() {
+			if ipnet.IP.To4() != nil {
+				return ipnet.IP.String()
+			}
+		}
+	}
+	return ""
+}
